Split and merge
