+++ config/splat.hd.ric.yaml
@@ -51,6 +51,7 @@ segments:
       - [0x1938C, .data, pl_anims]
       - [0x199AC, .data, 2A060]
       - [0x19D94, .data, 2C4C4]
+      - [0x19E98, .data, 2F8E8]
       - [0x19EE0, .data, 319C4]
       - [0x19EFC, .data, e_giant_spinning_cross]
       - [0x1A444, .rodata, pl_main]
@@ -64,6 +65,7 @@ segments:
       - [0x1A9B4, .rodata, 26C84]
       - [0x1A9F4, .rodata, 2A060]
       - [0x1AA88, .rodata, 2C4C4]
+      - [0x1AAF0, .rodata, 2F8E8]
       - [0x1AB04, .rodata, 319C4]
       - [0x1AC1C, .rodata, e_giant_spinning_cross]
       - [0x1AC2C, c, pl_main]
@@ -77,6 +79,7 @@ segments:
       - [0x26DD8, c, 26C84]
       - [0x29F9C, c, 2A060]
       - [0x2C42C, c, 2C4C4]
+      - [0x2F83C, c, 2F8E8]
       - [0x31918, c, 319C4]
       - [0x371E0, c, e_giant_spinning_cross]
       - [0x37AB8, .bss, pl_handlers]
@@ -84,6 +87,7 @@ segments:
       - [0x38ED4, .bss, 238F8]
       - [0x38F54, .bss, 2A060]
       - [0x38FDC, .bss, 2C4C4]
+      - [0x397E0, .bss, 2F8E8]
       - [0x397E8, .bss, 319C4]
       - [0x39834, bss]
   - [0x399AC]
