+++ config/splat.us.ric.yaml
@@ -50,6 +50,7 @@ segments:
       - [0x1938C, .data, pl_anims]
       - [0x199AC, .data, 2A060]
       - [0x19D94, .data, 2C4C4]
+      - [0x19E98, .data, 2F8E8]
       - [0x19EE0, .data, 319C4]
       - [0x19EFC, .data, e_giant_spinning_cross]
       - [0x1A444, .rodata, pl_main] # UpdateEntityRichter
@@ -71,6 +72,7 @@ segments:
       - [0x1AAEC, .rodata, 2C4C4] # RicEntityRevivalColumn
       - [0x1AB04, .rodata, 2C4C4] # RicEntitySubwpnCross
       - [0x1AB24, .rodata, 2C4C4] # RicEntitySubwpnAgunea
+      - [0x1AB24, .rodata, 2F8E8]
       - [0x1AB38, .rodata, 319C4]
       - [0x1AB68, .rodata, e_giant_spinning_cross]
       - [0x1AC50, .rodata, e_giant_spinning_cross]
@@ -85,6 +87,7 @@ segments:
       - [0x26C84, c, 26C84]
       - [0x2A060, c, 2A060]
       - [0x2C4C4, c, 2C4C4]
+      - [0x2F8E8, c, 2F8E8]
       - [0x319C4, c, 319C4]
       - [0x359A4, c, 319C4]
       - [0x3728C, c, e_giant_spinning_cross]
@@ -93,6 +96,7 @@ segments:
       - [0x37F84, .bss, 238F8]
       - [0x38000, .bss, 2A060]
       - [0x38088, .bss, 2C4C4]
+      - [0x3988C, .bss, 2F8E8]
       - [0x39894, .bss, 319C4]
       - [0x398E0, .bss, bss]
   - [0x39A58]
