+++ src/ric/2F8E8.c
@@ -1,0 +1,960 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "ric.h"
+#include "sfx.h"
+
+#include "../rebound_stone.h"
+
+// RIC entity #42. Blueprint 50. Comes from subweapon 7.
+// Rebound stone!
+void RicEntitySubwpnReboundStone(Entity* self) {
+    s16 playerX;
+    s16 playerY;
+    Collider collider;
+    s32 speed;
+    s32 currX;
+    s32 currY;
+    s32 collX;
+    s32 collY;
+    s32 deltaX;
+    s32 deltaY;
+    s32 i;
+    s32 colliderFlags;
+    PrimLineG2* prim;
+
+    speed = 0x400;
+    self->ext.reboundStone.unk82 = 0;
+    playerX = self->posX.i.hi;
+    playerY = self->posY.i.hi;
+
+    switch (self->step) {
+    case 0:
+        self->primIndex = g_api.AllocPrimitives(PRIM_LINE_G2, 16);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+        self->posY.i.hi -= 0x10;
+        playerY = self->posY.i.hi;
+        for (prim = (PrimLineG2*)&g_PrimBuf[self->primIndex], i = 0;
+             prim != NULL; i++, prim = prim->next) {
+            PGREY(prim, 0) = PGREY(prim, 1) = 0xFF;
+            prim->priority = PLAYER.zPriority + 2;
+            prim->drawMode =
+                DRAW_TPAGE2 | DRAW_TPAGE | DRAW_UNK02 | DRAW_TRANSP;
+            if (i != 0) {
+                prim->drawMode |= DRAW_HIDE;
+            }
+            prim->x0 = prim->x1 = playerX;
+            prim->y0 = prim->y1 = playerY;
+            prim->timer = 20;
+        }
+        self->flags =
+            FLAG_POS_CAMERA_LOCKED | FLAG_KEEP_ALIVE_OFFCAMERA | FLAG_HAS_PRIMS;
+        self->zPriority = PLAYER.zPriority + 2;
+
+        if (PLAYER.facingLeft) {
+            self->ext.reboundStone.stoneAngle = 0x980;
+        } else {
+            self->ext.reboundStone.stoneAngle = 0xE80;
+        }
+        self->ext.reboundStone.stoneAngle += (rand() & 0x7F) - 0x40;
+
+        self->ext.reboundStone.lifeTimer = 0x40;
+        self->ext.reboundStone.subweaponId = PL_W_REBNDSTONE;
+        RicSetSubweaponParams(self);
+        self->hitboxWidth = 4;
+        self->hitboxHeight = 4;
+        g_api.CheckCollision(self->posX.i.hi, self->posY.i.hi, &collider, 0);
+        if (collider.effects & EFFECT_SOLID) {
+            self->ext.reboundStone.unk84 = 4;
+        }
+        self->step++;
+        g_api.PlaySfx(SFX_WEAPON_SWISH_C);
+        break;
+    case 1:
+        deltaX = rcos(self->ext.reboundStone.stoneAngle) * 0x10;
+        deltaY = -rsin(self->ext.reboundStone.stoneAngle) * 0x10;
+        currX = self->posX.val;
+        currY = self->posY.val;
+        if (!self->ext.reboundStone.unk84) {
+            for (i = 0; i < 6; i++) {
+                collX = FIX_TO_I(currX);
+                collY = FIX_TO_I(currY + deltaY);
+                g_api.CheckCollision(collX, collY, &collider, 0);
+                colliderFlags =
+                    collider.effects &
+                    (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_2000 |
+                     EFFECT_UNK_1000 | EFFECT_UNK_0800 | EFFECT_UNK_0002 |
+                     EFFECT_SOLID);
+                if (colliderFlags & EFFECT_SOLID) {
+                    colliderFlags &= 0xFF00;
+                    if (deltaY > 0) {
+                        if ((colliderFlags == 0) ||
+                            (colliderFlags & EFFECT_UNK_0800)) {
+                            ReboundStoneBounce1(0x800);
+                        }
+                        if (colliderFlags == EFFECT_UNK_8000) {
+                            ReboundStoneBounce2(0x200);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_8000 + EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0x12E);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_8000 + EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0xA0);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_8000 + EFFECT_UNK_4000) {
+                            ReboundStoneBounce2(0x600);
+                        }
+                        if (colliderFlags == EFFECT_UNK_8000 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0x6D2);
+                        }
+                        if (colliderFlags == EFFECT_UNK_8000 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0x760);
+                        }
+                    }
+                    if (deltaY < 0) {
+                        if ((colliderFlags == 0) ||
+                            (colliderFlags & EFFECT_UNK_8000)) {
+                            ReboundStoneBounce1(0x800);
+                        }
+                        if (colliderFlags == EFFECT_UNK_0800) {
+                            ReboundStoneBounce2(0xE00);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_0800 + EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0xED2);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_0800 + EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0xF60);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_0800 + EFFECT_UNK_4000) {
+                            ReboundStoneBounce2(0xA00);
+                        }
+                        if (colliderFlags == EFFECT_UNK_0800 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0x92E);
+                        }
+                        if (colliderFlags == EFFECT_UNK_0800 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0x8A0);
+                        }
+                    }
+                }
+                collY = FIX_TO_I(currY);
+                collX = FIX_TO_I(currX + deltaX);
+                g_api.CheckCollision(collX, collY, &collider, 0);
+                colliderFlags =
+                    collider.effects &
+                    (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_2000 |
+                     EFFECT_UNK_1000 | EFFECT_UNK_0800 | EFFECT_UNK_0002 |
+                     EFFECT_SOLID);
+                if (colliderFlags & EFFECT_SOLID) {
+                    colliderFlags &= 0xFF00;
+                    // Cases when traveling right
+                    if (deltaX > 0) {
+                        if ((colliderFlags == 0) ||
+                            TEST_BITS(colliderFlags, 0x4800) ||
+                            TEST_BITS(colliderFlags, 0xC000)) {
+                            ReboundStoneBounce1(0x400);
+                        }
+                        if (colliderFlags == EFFECT_UNK_0800) {
+                            ReboundStoneBounce2(0xE00);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_0800 + EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0xED2);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_0800 + EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0xF60);
+                        }
+                        if (colliderFlags == EFFECT_UNK_8000) {
+                            ReboundStoneBounce2(0x200);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_8000 + EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0x12E);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_8000 + EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0xA0);
+                        }
+                    }
+                    // Cases when traveling left
+                    if (deltaX < 0) {
+                        if ((colliderFlags == 0) ||
+                            ((colliderFlags & 0x4800) == 0x800) ||
+                            ((colliderFlags & 0xC000) == 0x8000)) {
+                            ReboundStoneBounce1(0x400);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_0800 + EFFECT_UNK_4000) {
+                            ReboundStoneBounce2(0xA00);
+                        }
+                        if (colliderFlags == EFFECT_UNK_0800 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0x92E);
+                        }
+                        if (colliderFlags == EFFECT_UNK_0800 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0x8A0);
+                        }
+                        if (colliderFlags ==
+                            EFFECT_UNK_8000 + EFFECT_UNK_4000) {
+                            ReboundStoneBounce2(0x600);
+                        }
+                        if (colliderFlags == EFFECT_UNK_8000 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_1000) {
+                            ReboundStoneBounce2(0x6D2);
+                        }
+                        if (colliderFlags == EFFECT_UNK_8000 + EFFECT_UNK_4000 +
+                                                 EFFECT_UNK_2000) {
+                            ReboundStoneBounce2(0x760);
+                        }
+                    }
+                }
+                if (self->ext.reboundStone.unk82) {
+                    goto block_93;
+                }
+                currX += deltaX;
+                currY += deltaY;
+            }
+        } else {
+            self->ext.reboundStone.unk84--;
+        }
+    block_93:
+        if (self->ext.reboundStone.unk82) {
+            g_api.CreateEntFactoryFromEntity(self, FACTORY(BP_42, 2), 0);
+            g_api.PlaySfx(SFX_UI_TINK);
+        }
+        if (self->posX.i.hi < -0x40 || self->posX.i.hi > 0x140 ||
+            self->posY.i.hi < -0x40 || self->posY.i.hi > 0x140 ||
+            self->ext.reboundStone.unk80 == 15) {
+            self->step = 2;
+        } else {
+            deltaX =
+                ((rcos(self->ext.reboundStone.stoneAngle) << 4) * speed) >> 8;
+            self->posX.val += deltaX;
+            deltaY =
+                -((rsin(self->ext.reboundStone.stoneAngle) << 4) * speed) >> 8;
+            self->posY.val += deltaY;
+        }
+        break;
+    case 2:
+        if (--self->ext.reboundStone.lifeTimer == 0) {
+            DestroyEntity(self);
+            return;
+        }
+        if (self->ext.reboundStone.lifeTimer == 0x20) {
+            self->hitboxState = 0;
+        }
+        prim = (PrimLineG2*)&g_PrimBuf[self->primIndex];
+        while (prim != NULL) {
+            prim->timer = 0;
+            prim = prim->next;
+        }
+        break;
+    }
+
+    prim = (PrimLineG2*)&g_PrimBuf[self->primIndex];
+    i = 0;
+    if (self->step == 2) {
+        colliderFlags = 4;
+    } else {
+        colliderFlags = 2;
+    }
+    // cleaner to use previous 3 lines than to put them in the for's initializer
+    for (; prim != NULL; i++, prim = prim->next) {
+        if (self->ext.reboundStone.unk82) {
+            if (i == self->ext.reboundStone.unk80) {
+                prim->x0 = playerX;
+                prim->y0 = playerY;
+                prim->drawMode &= ~DRAW_HIDE;
+            }
+        }
+        if (i == self->ext.reboundStone.unk80) {
+            prim->x1 = self->posX.i.hi;
+            prim->y1 = self->posY.i.hi;
+        }
+        if (!(prim->drawMode & DRAW_HIDE)) {
+            if (prim->timer) {
+                prim->timer--;
+            } else {
+                // again not colliderFlags, seems to control stone fading
+                if (colliderFlags < prim->b1) {
+                    prim->b1 -= colliderFlags;
+                }
+                PGREY(prim, 0) = PGREY(prim, 1);
+            }
+        }
+    }
+}
+
+// RIC entity #43. Blueprint 49. Subweapon 8. Vibhuti!
+void RicEntitySubwpnThrownVibhuti(Entity* self) {
+    Collider col;
+    FakePrim* prim;
+    s16 randomAngle;
+    s16 x;
+    s16 y;
+    s16 temp; // used for multiple unrelated things
+    s32 i;
+
+    switch (self->step) {
+    case 0:
+        self->primIndex = g_api.func_800EDB58(PRIM_TILE_ALT, 13);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+        self->flags = FLAG_POS_CAMERA_LOCKED | FLAG_HAS_PRIMS;
+        self->ext.subweapon.subweaponId = PL_W_VIBHUTI;
+        RicSetSubweaponParams(self);
+        self->hitboxWidth = self->hitboxHeight = 4;
+        self->ext.subweapon.timer = 0x80;
+        prim = (FakePrim*)&g_PrimBuf[self->primIndex];
+        self->posY.i.hi -= 15;
+        if (PLAYER.facingLeft) {
+            self->posX.i.hi -= 13;
+        } else {
+            self->posX.i.hi += 13;
+        }
+        x = self->posX.i.hi;
+        y = self->posY.i.hi;
+        i = 0;
+        while (true) {
+            prim->drawMode = DRAW_UNK02;
+            prim->priority = PLAYER.zPriority - 1;
+            if (prim->next == NULL) {
+                prim->drawMode &= ~DRAW_HIDE;
+                prim->y0 = prim->x0 = prim->w = 0;
+                break;
+            }
+            prim->posX.i.hi = x;
+            prim->posY.i.hi = y;
+            prim->posX.i.lo = prim->posY.i.lo = 0;
+            randomAngle = (rand() & 0xFF) + 0x100;
+            temp = (rand() & 0xFF) + 0x80;
+            prim->velocityX.val = ((rcos(randomAngle) << 4) * temp >> 9);
+            prim->velocityX.val += FIX(0.5);
+            prim->velocityY.val = -((rsin(randomAngle) << 4) * temp >> 9);
+            prim->velocityX.val = (prim->velocityX.val * 3) >> 1;
+            if (self->facingLeft) {
+                prim->velocityX.val = -prim->velocityX.val;
+            }
+            prim->posY.i.hi -= 4;
+            prim->delay = 1;
+            prim->x0 = prim->posX.i.hi;
+            prim->y0 = prim->posY.i.hi;
+            prim->r0 = 0xFF;
+            prim->g0 = 0xFF;
+            prim->b0 = 0xFF;
+            prim->w = 2;
+            prim->h = 2;
+            i++;
+            prim = prim->next;
+        }
+        g_api.PlaySfx(SFX_WEAPON_SWISH_C);
+        self->step++;
+        break;
+    case 1:
+        temp = 2;
+        if (self->facingLeft) {
+            temp = -temp;
+        }
+        if (--self->ext.subweapon.timer == 0) {
+            DestroyEntity(self);
+            return;
+        }
+        prim = (FakePrim*)&g_PrimBuf[self->primIndex];
+        i = 0;
+        while (true) {
+            if (prim->next == NULL) {
+                prim->drawMode &= ~DRAW_HIDE;
+                prim->y0 = prim->x0 = prim->w = 0;
+                break;
+            }
+            prim->posX.i.hi = prim->x0;
+            prim->posY.i.hi = prim->y0;
+            if (prim->delay) {
+                if (prim->velocityX.val != 0) {
+                    prim->posX.val += prim->velocityX.val;
+                    g_api.CheckCollision(
+                        prim->posX.i.hi + temp, prim->posY.i.hi, &col, 0);
+                    if (col.effects & EFFECT_UNK_0002) {
+                        prim->velocityX.val = 0;
+                    }
+                }
+                prim->posY.val += prim->velocityY.val;
+                prim->velocityY.val += FIX(12.0 / 128);
+                if (prim->velocityY.val > FIX(4)) {
+                    prim->velocityY.val = FIX(4);
+                }
+                if (prim->velocityY.val > 0) {
+                    g_api.CheckCollision(
+                        prim->posX.i.hi, prim->posY.i.hi, &col, 0);
+                    if (col.effects & EFFECT_SOLID) {
+                        prim->delay = 0;
+                        prim->posY.i.hi += col.unk18;
+                        prim->posY.i.hi -= (i % 3 + 1);
+                        prim->w = prim->h = 3;
+                    }
+                }
+            }
+            if ((self->ext.subweapon.timer & 7) == i) {
+                self->posX.i.hi = prim->posX.i.hi;
+                self->posY.i.hi = prim->posY.i.hi;
+                if (prim->drawMode & DRAW_HIDE) {
+                    self->hitboxWidth = self->hitboxHeight = 0;
+                } else {
+                    self->hitboxWidth = self->hitboxHeight = 4;
+                }
+                if (prim->delay) {
+                    self->hitboxOffY = 0;
+                } else {
+                    self->hitboxOffY = -6;
+                }
+            }
+            if (self->hitFlags && ((self->ext.subweapon.timer + 1) & 7) == i) {
+                prim->drawMode = DRAW_HIDE;
+            }
+            if (self->ext.subweapon.timer - 1 == i) {
+                prim->drawMode = DRAW_HIDE;
+            }
+            i++;
+            prim->x0 = prim->posX.i.hi;
+            prim->y0 = prim->posY.i.hi;
+            prim = prim->next;
+        }
+        self->hitFlags = 0;
+        break;
+    }
+}
+
+static u8 PrimDecreaseBrightness(Primitive* prim, u8 amount) {
+    s32 i;
+    s32 j;
+    u8* colorPtr;   // points to an RGB color
+    u8* channelPtr; // points to a single channel of that color
+    u8 isEnd;
+
+    isEnd = 0;
+    colorPtr = &prim->r0;
+    for (i = 0; i < 4; colorPtr += OFF(Primitive, r1) - OFF(Primitive, r0),
+        i++) {
+        for (j = 0; j < 3; j++) {
+            channelPtr =
+                &colorPtr[j]; // get the red, green, blue, or pad channel
+            *channelPtr -= amount;
+
+            if (*channelPtr < 16) {
+                *channelPtr = 16;
+            } else {
+                isEnd |= 1;
+            }
+        }
+    }
+    return isEnd;
+}
+
+// RIC Entity #44. Blueprint 51. Subweapon 9. Agunea (non-crash). DRA
+// blueprint 25. DRA entity 21.
+void RicEntitySubwpnAgunea(Entity* self) {
+    Primitive* prim;
+    s32 heartCost;
+    s16 tempX;
+    s16 tempY;
+    u32 heartBroachesWorn;
+
+    if (g_Player.status & (PLAYER_STATUS_TRANSFORM | PLAYER_STATUS_UNK10000)) {
+        DestroyEntity(self);
+        return;
+    }
+    switch (self->step) {
+    case 0:
+        self->primIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+        self->flags =
+            FLAG_POS_CAMERA_LOCKED | FLAG_KEEP_ALIVE_OFFCAMERA | FLAG_HAS_PRIMS;
+        self->facingLeft = PLAYER.facingLeft;
+        self->ext.agunea.subweaponId = PL_W_AGUNEA;
+        RicSetSubweaponParams(self);
+        self->hitboxWidth = self->hitboxHeight = 4;
+        self->hitboxOffX = 4;
+        self->hitboxOffY = 0;
+        self->posY.i.hi = self->ext.agunea.unk82 =
+            PLAYER.posY.i.hi + PLAYER.hitboxOffY - 8;
+        self->posX.i.hi = self->ext.agunea.unk80 = PLAYER.posX.i.hi;
+        prim = &g_PrimBuf[self->primIndex];
+        prim->type = PRIM_LINE_G2;
+        prim->priority = PLAYER.zPriority + 2;
+        prim->drawMode = DRAW_UNK_200 | DRAW_UNK_100 | DRAW_TPAGE2 |
+                         DRAW_TPAGE | DRAW_TRANSP;
+        prim->r1 = 0x60;
+        prim->g1 = 0;
+        prim->b1 = 0x80;
+        RicSetSpeedX(FIX(6));
+        g_api.PlaySfx(SFX_WEAPON_SWISH_C);
+        self->step++;
+        break;
+    case 1:
+        self->posX.val += self->velocityX;
+        if (self->posX.i.hi < -0x40 || self->posX.i.hi > 0x140 ||
+            self->posY.i.hi < -0x20 || self->posY.i.hi > 0x120) {
+            self->step = 2;
+        }
+        if (self->hitFlags) {
+            self->ext.agunea.parent = self->unkB8;
+            self->step = 3;
+        }
+        break;
+    case 4:
+        self->posX.i.hi = self->ext.agunea.parent->posX.i.hi;
+        self->posY.i.hi = self->ext.agunea.parent->posY.i.hi;
+        if (++self->ext.agunea.unk7C < 16) {
+            break;
+        }
+        // Fallthrough
+    case 2:
+        prim = &g_PrimBuf[self->primIndex];
+        if (prim->r1 < 5) {
+            DestroyEntity(self);
+            return;
+        }
+        break;
+    case 3:
+        if ((g_Player.padPressed & (PAD_UP + PAD_SQUARE)) !=
+            (PAD_UP + PAD_SQUARE)) {
+            self->step = 4;
+        }
+        if (!self->ext.agunea.parent->entityId) {
+            self->step = 2;
+            return;
+        }
+        if (self->ext.agunea.unk7C &&
+            (self->ext.agunea.parent->hitPoints > 0x7000 ||
+             !self->ext.agunea.parent->hitPoints ||
+             !self->ext.agunea.parent->hitboxState)) {
+            self->step = 2;
+            return;
+        }
+
+        tempX = self->posX.i.hi = self->ext.agunea.parent->posX.i.hi;
+        tempY = self->posY.i.hi = self->ext.agunea.parent->posY.i.hi;
+        if ((self->ext.agunea.unk7C % 12) == 0) {
+            self->posX.i.hi += (rand() & 0xF) - 8;
+            self->posY.i.hi += (rand() & 0xF) - 8;
+            if (g_Status.hearts >= 5) {
+                g_Status.hearts -= 5;
+                RicCreateEntFactoryFromEntity(self, BP_AGUNEA_HIT_ENEMY, 0);
+                g_api.PlaySfx(SFX_THUNDER_B);
+            } else if (self->ext.agunea.unk84 == 0) {
+                RicCreateEntFactoryFromEntity(self, BP_AGUNEA_HIT_ENEMY, 0);
+                g_api.PlaySfx(SFX_THUNDER_B);
+                self->ext.agunea.unk84++;
+            } else {
+                self->step = 4;
+            }
+        }
+        self->posX.i.hi = tempX;
+        self->posY.i.hi = tempY;
+        self->ext.agunea.unk7C++;
+        break;
+    }
+    prim = &g_PrimBuf[self->primIndex];
+    if (prim->r1 >= 4) {
+        prim->r1 -= 4;
+    }
+    if (prim->g1 >= 4) {
+        prim->g1 -= 4;
+    }
+    if (prim->b1 >= 4) {
+        prim->b1 -= 4;
+    }
+    if (prim->b1 < 5) {
+        prim->drawMode |= DRAW_HIDE;
+    }
+    prim->x0 = self->ext.agunea.unk80;
+    prim->y0 = self->ext.agunea.unk82;
+    prim->x1 = self->posX.i.hi;
+    prim->y1 = self->posY.i.hi;
+}
+
+void RicEntityAguneaHitEnemy(Entity* self) {
+    Entity* parent;
+    Primitive* prim;
+    Primitive* temp_s3;
+    Primitive* var_a0;
+    s16 arctan;
+    s16 angle;
+    s16 xOffset;
+    s16 yOffset;
+    s16 temp_s2;
+    u8 var_s3;
+    s32 i;
+    u8 var_s8;
+
+    parent = self->ext.et_801291C4.parent;
+    self->posX.i.hi = PLAYER.posX.i.hi;
+    self->posY.i.hi = (PLAYER.posY.i.hi + PLAYER.hitboxOffY) - 8;
+    if (self->ext.et_801291C4.parent->entityId != 0x2C) {
+        switch (self->step) {
+        case 0:
+            DestroyEntity(self);
+            return;
+        case 1:
+        case 2:
+        case 4:
+            self->step = 3;
+        }
+    }
+    switch (self->step) {
+    case 0:
+        self->primIndex = g_api.AllocPrimitives(PRIM_GT4, 0x28);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            break;
+        }
+        self->flags = FLAG_KEEP_ALIVE_OFFCAMERA | FLAG_HAS_PRIMS;
+        self->facingLeft = PLAYER.facingLeft;
+        self->ext.et_801291C4.unk84 = (rand() & 0x3FF) - 0x200;
+        if (self->facingLeft) {
+            // @bug: This should be assigned to something. As-is, does nothing.
+            self->ext.et_801291C4.unk84 + 0x800;
+        }
+        self->ext.et_801291C4.unk84 &= 0xFFF;
+        self->ext.et_801291C4.unk90 = (self->params >> 8) & 0xFF;
+        prim = &g_PrimBuf[self->primIndex];
+        self->ext.et_801291C4.prim1 = prim;
+        self->ext.et_801291C4.prim2 = prim;
+        for (i = 0; prim != NULL;) {
+            prim->tpage = 0x1A;
+            prim->clut = 0x194;
+            prim->u0 = prim->u1 = i * 0x10 + 0x90;
+            prim->u2 = prim->u3 = prim->u0 + 0x10;
+            prim->v0 = prim->v2 = 0xD0;
+            prim->v1 = prim->v3 = 0xC0;
+            prim->x0 = self->posX.i.hi;
+            prim->y0 = self->posY.i.hi;
+            prim->x2 = self->posX.i.hi;
+            prim->y2 = self->posX.i.hi;
+            prim->r0 = prim->g0 = prim->b0 = 0xF0;
+            LOW(prim->r1) = LOW(prim->r0);
+            LOW(prim->r2) = LOW(prim->r0);
+            LOW(prim->r3) = LOW(prim->r0);
+            prim->priority = self->zPriority;
+            prim->drawMode = DRAW_HIDE;
+            prim = prim->next;
+            i++;
+            if (i > 5) {
+                i = 0;
+            }
+        }
+        prim = self->ext.et_801291C4.prim1;
+        prim->x0 = self->posX.i.hi;
+        prim->y0 = self->posY.i.hi;
+        prim->x1 = prim->x0;
+        prim->y1 = prim->y1 - 0x10;
+        prim->x2 = self->posX.i.hi;
+        prim->y2 = self->posY.i.hi;
+        prim->x3 = prim->x2;
+        prim->y3 = prim->y2 - 0x10;
+        self->ext.et_801291C4.prim2 = prim;
+        while (prim != NULL) {
+            prim->clut = 0x194;
+            prim->r0 = prim->g0 = prim->b0 = 0x80;
+            LOW(prim->r1) = LOW(prim->r0);
+            LOW(prim->r2) = LOW(prim->r0);
+            LOW(prim->r3) = LOW(prim->r0);
+            prim->priority = self->zPriority;
+            prim->drawMode = DRAW_HIDE;
+            prim = prim->next;
+        }
+        self->ext.et_801291C4.unk88 = 0;
+        self->step++;
+        break;
+    case 1:
+        for (i = 0; i < 2; i++) {
+            prim = self->ext.et_801291C4.prim2;
+            temp_s2 = self->ext.et_801291C4.unk84;
+            xOffset = parent->posX.i.hi - prim->x2;
+            yOffset = parent->posY.i.hi - prim->y2;
+            if (abs(xOffset) < 8 && abs(yOffset) < 8) {
+                self->step++;
+                break;
+            }
+            if (abs(xOffset) < 0x40 && abs(yOffset) < 0x40) {
+                var_s3 = 1;
+            } else {
+                var_s3 = 0;
+            }
+            if (!self->ext.et_801291C4.unk88) {
+                self->ext.et_801291C4.unk88 = 4;
+                if (var_s3) {
+                    self->ext.et_801291C4.unk88 = 2;
+                }
+                arctan = ratan2(-yOffset, xOffset);
+                angle = arctan - temp_s2;
+                if (angle > 0x800) {
+                    angle = 0x1000 - angle;
+                }
+                if (angle < -0x800) {
+                    angle = 0x1000 + angle;
+                }
+                if (!var_s3) {
+                    angle /= 4;
+                } else {
+                    angle /= 2;
+                }
+                self->ext.et_801291C4.unk86 = angle;
+            }
+            temp_s2 += self->ext.et_801291C4.unk86;
+            if (!var_s3) {
+                temp_s2 += 0x180 - ((rand() & 3) << 8);
+            }
+            temp_s2 &= 0xFFF;
+            temp_s3 = prim->next;
+            if (temp_s3 == NULL) {
+                self->step++;
+                return;
+            }
+            LOW(temp_s3->x0) = LOW(prim->x2);
+            LOW(temp_s3->x1) = LOW(prim->x3);
+            self->ext.et_801291C4.unk84 = temp_s2;
+            self->ext.et_801291C4.prim2 = temp_s3;
+            xOffset = (rcos(temp_s2) * 0xC) >> 0xC;
+            yOffset = -((rsin(temp_s2) * 0xC) >> 0xC);
+            temp_s3->x2 = temp_s3->x0 + xOffset;
+            temp_s3->y2 = temp_s3->y0 + yOffset;
+            angle = temp_s2 - 0x400;
+            var_s8 = 0x10 - (self->params * 4);
+            xOffset = (var_s8 * rcos(angle)) >> 0xC;
+            yOffset = -((var_s8 * rsin(angle)) >> 0xC);
+            temp_s3->x3 = temp_s3->x2 + xOffset;
+            temp_s3->y3 = temp_s3->y2 + yOffset;
+            temp_s3->drawMode = DRAW_COLORS | DRAW_UNK02;
+            self->ext.et_801291C4.unk88--;
+        }
+        break;
+    case 2:
+        if (!self->step_s) {
+            prim = self->ext.et_801291C4.prim1;
+            while (prim != NULL) {
+                prim->clut = 0x15F;
+                prim = prim->next;
+            }
+            self->step_s++;
+            return;
+        }
+        prim = self->ext.et_801291C4.prim1;
+        while (prim != NULL) {
+            prim->clut = 0x194;
+            prim->r0 = prim->g0 = prim->b0 = 0x60;
+            LOW(prim->r1) = LOW(prim->r0);
+            LOW(prim->r2) = LOW(prim->r0);
+            LOW(prim->r3) = LOW(prim->r0);
+            prim = prim->next;
+        }
+        self->step_s = 0;
+        self->step++;
+        break;
+    case 3:
+        var_s8 = 1;
+        prim = self->ext.et_801291C4.prim1;
+        while (prim != NULL) {
+            var_s8 &= !PrimDecreaseBrightness(prim, 4);
+            prim = prim->next;
+        }
+        if (var_s8) {
+            prim = self->ext.et_801291C4.prim1;
+            while (prim != NULL) {
+                prim->drawMode = DRAW_HIDE;
+                prim = prim->next;
+            }
+            DestroyEntity(self);
+        }
+        break;
+    }
+}
+
+// White cloud that comes out of Richter's finger with the Vibhuti crash
+static AnimationFrame D_80155EA8[] = {
+    {3, FRAME(1, 0)},  {3, FRAME(2, 0)},  {3, FRAME(3, 0)},  {3, FRAME(4, 0)},
+    {3, FRAME(5, 0)},  {3, FRAME(6, 0)},  {3, FRAME(7, 0)},  {3, FRAME(8, 0)},
+    {3, FRAME(9, 0)},  {3, FRAME(10, 0)}, {3, FRAME(11, 0)}, {3, FRAME(12, 0)},
+    {3, FRAME(13, 0)}, A_LOOP_AT(0)};
+void RicEntityVibhutiCrashCloud(Entity* entity) {
+    s32 angle;
+
+    switch (entity->step) {
+    case 0:
+        entity->primIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
+        if (entity->primIndex == -1) {
+            DestroyEntity(entity);
+            return;
+        }
+        entity->flags = FLAG_POS_CAMERA_LOCKED | FLAG_HAS_PRIMS;
+        entity->posX.val = entity->ext.vibCrashCloud.parent->ext.vibhutiCrash.x;
+        entity->posY.val = entity->ext.vibCrashCloud.parent->ext.vibhutiCrash.y;
+        entity->facingLeft =
+            entity->ext.vibCrashCloud.parent->ext.vibhutiCrash.facing;
+        entity->ext.vibCrashCloud.subweaponId = PL_W_CRASH_VIBHUTI;
+        RicSetSubweaponParams(entity);
+        entity->flags |= FLAG_UNK_100000;
+        entity->unk5A = 0x79;
+        entity->animSet = ANIMSET_DRA(14);
+        entity->palette = PAL_OVL(0x19E);
+        entity->anim = D_80155EA8;
+        entity->drawMode = DRAW_TPAGE2 | DRAW_TPAGE;
+        entity->drawFlags = FLAG_DRAW_UNK8;
+        entity->unk6C = 0x60;
+        entity->hitboxWidth = 8;
+        entity->hitboxHeight = 8;
+        angle = (rand() % 512) + 0x300;
+        entity->velocityX = rcos(angle) * 32;
+        entity->velocityY = -(rsin(angle) * 32);
+        entity->step++;
+        break;
+    case 1:
+        entity->ext.vibCrashCloud.unk7C++;
+        if (entity->ext.vibCrashCloud.unk7C >= 39) {
+            DestroyEntity(entity);
+        } else {
+            entity->posX.val += entity->velocityX;
+            entity->posY.val += entity->velocityY;
+        }
+        break;
+    }
+}
+
+#if defined(VERSION_PSP)
+extern s32 crash_vibhuti_timer;
+#else
+static s32 crash_vibhuti_timer;
+#endif
+void RicEntityCrashVibhuti(Entity* self) {
+    FakePrim* prim;
+    s32 angle;
+    s32 magnitude;
+    s32 i;
+    s32 facing;
+
+    switch (self->step) {
+    case 0:
+        self->primIndex = g_api.AllocPrimitives(PRIM_TILE, 9);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            g_Player.unk4E = 1;
+            return;
+        }
+        self->flags =
+            FLAG_KEEP_ALIVE_OFFCAMERA | FLAG_HAS_PRIMS | FLAG_UNK_20000;
+        prim = (FakePrim*)&g_PrimBuf[self->primIndex];
+        for (i = 0; i < 9; i++) {
+            prim->r0 = prim->g0 = prim->b0 = 0xFF;
+            prim->w = prim->h = 1;
+            prim->priority = PLAYER.zPriority + 8;
+            prim->drawMode = DRAW_HIDE | DRAW_UNK02;
+            prim = prim->next;
+        }
+        crash_vibhuti_timer = 0;
+        self->step++;
+        break;
+    case 1:
+        // Weird fake stuff to load unk7E a second time
+        self->ext.vibhutiCrash.unk7E++;
+        if (!(self->ext.vibhutiCrash.unk7E % 2) &&
+            self->ext.vibhutiCrash.timer < 8) {
+            self->ext.vibhutiCrash.timer++;
+            self->ext.vibhutiCrash.unk80++;
+            if (self->ext.vibhutiCrash.unk80 >= 0x30) {
+                self->step++;
+            }
+            prim = (FakePrim*)&g_PrimBuf[self->primIndex];
+            for (i = 0; i < 9; i++) {
+                if (prim->drawMode & DRAW_HIDE) {
+                    break;
+                }
+                prim = prim->next;
+            }
+            prim->posX.val = PLAYER.posX.val;
+            prim->posY.val = PLAYER.posY.val + FIX(-24);
+            angle = rand() % 0x200 + 0x300;
+            magnitude = (rand() % 24) + 0x20;
+            prim->velocityX.val = (rcos(angle) * magnitude);
+            prim->velocityY.val = -(rsin(angle) * magnitude);
+            prim->drawMode &= ~DRAW_HIDE;
+            prim->delay = 0x10;
+        }
+        // fallthrough
+    case 2:
+        crash_vibhuti_timer++;
+        if (!(crash_vibhuti_timer % 8)) {
+            g_api.PlaySfx(SFX_NOISE_SWEEP_DOWN_A);
+        }
+        prim = (FakePrim*)&g_PrimBuf[self->primIndex];
+        for (i = 0; i < 9; i++) {
+            if (!(prim->drawMode & DRAW_HIDE)) {
+                if (!--prim->delay) {
+                    prim->drawMode |= DRAW_HIDE;
+                    self->ext.vibhutiCrash.timer--;
+                    self->ext.vibhutiCrash.x = prim->posX.val;
+                    self->ext.vibhutiCrash.y = prim->posY.val;
+                    if (prim->velocityX.val > 0) {
+                        facing = false;
+                    } else {
+                        facing = true;
+                    }
+                    self->ext.vibhutiCrash.facing = facing;
+                    RicCreateEntFactoryFromEntity(
+                        self, BP_VITHUBI_CRASH_CLOUD, 0);
+                } else {
+                    prim->posX.val += prim->velocityX.val;
+                    prim->posY.val += prim->velocityY.val;
+                    prim->velocityY.val += FIX(0.25);
+                    prim->x0 = prim->posX.i.hi;
+                    prim->y0 = prim->posY.i.hi;
+                }
+            }
+            prim = prim->next;
+        }
+        if (self->step == 2 && !self->ext.vibhutiCrash.timer) {
+            self->step++;
+        }
+        break;
+    case 3:
+        g_Player.unk4E = 1;
+        g_Player.unk4E = 1;
+        DestroyEntity(self);
+        break;
+    }
+}
+
+void RicEntityCrashReboundStoneParticles(Entity* entity) {
+    switch (entity->step) {
+    case 0:
+        entity->flags = FLAG_KEEP_ALIVE_OFFCAMERA;
+        entity->ext.subweapon.subweaponId = PL_W_CRASH_REBOUND_STONE;
+        RicSetSubweaponParams(entity);
+        entity->hitboxWidth = 4;
+        entity->hitboxHeight = 4;
+        entity->step++;
+        break;
+    case 1:
+        entity->ext.subweapon.timer++;
+        if (entity->ext.subweapon.timer > 3) {
+            DestroyEntity(entity);
+        }
+        break;
+    }
+}
